refactor(navigation)-Extract screen transition logic
